+++ src/modules/ingredient/ingredient.service.ts
@@ -5,21 +5,34 @@ import { Model } from 'mongoose';
 import { IngredientDocument } from './ingredient.interface';
 import { CreateIngredientDto } from './ingredient.dto';
 import { LoggerService } from '../logger/logger.service';
+import { RedisService } from '../redis/redis.service';
 
 @Injectable()
 export class IngredientService {
 
     constructor(
-        @InjectModel(models.INGREDIENT_MODEL)
-        private ingredientModel: Model<IngredientDocument>,
-        private loggerService: LoggerService
+        @InjectModel(models.INGREDIENT_MODEL) private ingredientModel: Model<IngredientDocument>,
+        private readonly redisService: RedisService,
+        private readonly loggerService: LoggerService
     ) {}
 
     async findAll(): Promise<IngredientDocument[]> {
-        const ingredients = (await this.ingredientModel.find()) as IngredientDocument[];
-        const message = `Found ${ingredients.length} meals.`;
+        const cachedIngredients = await this.redisService.get<IngredientDocument>('ingredients');
+        const context = 'IngredientService/findAll';
 
-        this.loggerService.info('IngredientService/findAll:', message);
+        if (cachedIngredients) {
+            this.loggerService.info(context, `Fetched ${cachedIngredients.length} ingredients from cache.`);
+
+            return cachedIngredients;
+        }
+
+        const ingredients = (await this.ingredientModel.find()) as IngredientDocument[];
+        const message = `Found ${ingredients.length} ingredients.`;
+
+        await this.redisService.set<IngredientDocument>(ingredients, 'ingredients');
+        this.loggerService.info(context, 'Cached found ingredients.');
+
+        this.loggerService.info(context, message);
 
         return ingredients;
     }
@@ -28,7 +41,7 @@ export class IngredientService {
         const createdIngredient = await this.ingredientModel.create(createIngredientDto) as IngredientDocument;
         const message = `New ingredient "${createIngredientDto.name}" has been added.`;
 
-        this.loggerService.info('IngredientService/create:', message);
+        this.loggerService.info('IngredientService/create', message);
 
         return createdIngredient;
     }