+++ src/modules/redis/redis.constants.ts
@@ -1,1 +1,3 @@
 export const REDIS_CLIENT = 'REDIS_CLIENT';
+
+export const REDIS_TTL = 60 * 60; // 1 hour